Move auth and user middlewars to api middlewares dir
